refactor: moved io functions
